Rename leader election config package/methods
